Add "web-bot-auth/crypto" to ease signing
Instead of letting end-user define their own signing method every time,
we provide webcrypto based signers.

This commit adds one for Ed25519
