+++ packages/web-bot-auth/package.json
@@ -10,10 +10,15 @@
       "types": "./dist/index.d.ts",
       "require": "./dist/index.js",
       "import": "./dist/index.mjs"
+    },
+    "./crypto": {
+      "types": "./dist/crypto.d.ts",
+      "require": "./dist/crypto.js",
+      "import": "./dist/crypto.mjs"
     }
   },
   "scripts": {
-    "build": "tsup src/index.ts --format cjs,esm --dts --clean",
+    "build": "tsup src/index.ts src/crypto.ts --format cjs,esm --dts --clean",
     "generate-test-vectors": "node --experimental-transform-types scripts/test-vectors.ts test/test_data/web_bot_auth_architecture_v1.json",
     "prepublishOnly": "npm run build",
     "test": "vitest",
