+++ packages/web-bot-auth/test/test_data/web_bot_auth_architecture_v1.json
